requirements.txt

fastapi
uvicorn[standard]
torch
torchvision
pillow
python-multipart
jinja2
